+++ src/Controller/AppController.php
@@ -14,18 +14,19 @@
  */
 namespace App\Controller;
 
-use Cake\Controller\Controller;
+use DateTime;
+use Cake\Log\Log;
+use Cake\Cache\Cache;
 use Cake\Event\Event;
 use Cake\Core\Configure;
-use Cake\ORM\TableRegistry;
 use Cake\Routing\Router;
-use ControllerAction\Model\Traits\ControllerActionTrait;
-use ControllerAction\Model\Traits\SecurityTrait;
-use Cake\Utility\Inflector;
-use Cake\Cache\Cache;
 use Cake\Filesystem\File;
 use Cake\Filesystem\Folder;
-use Cake\Log\Log;
+use Cake\ORM\TableRegistry;
+use Cake\Utility\Inflector;
+use Cake\Controller\Controller;
+use ControllerAction\Model\Traits\SecurityTrait;
+use ControllerAction\Model\Traits\ControllerActionTrait;
 
 /**
  * Application Controller
@@ -258,6 +259,38 @@ class AppController extends Controller
             in_array($this->response->type(), ['application/json', 'application/xml'])
         ) {
             $this->set('_serialize', true);
+        }
+    }
+    
+    public function beforeFilter(Event $event)
+    {
+        $user = $this->Auth->user();
+        if(!is_null($user)){
+            $userData = TableRegistry::get("security_users")->get($user['id']);
+        try {
+            date_default_timezone_set('Asia/Colombo');
+            $start_date = new DateTime($userData->security_timeout);
+            $since_start = $start_date->diff(new DateTime(date('Y-m-d h:i:s a', time())));
+            $totMonths = $since_start->y *12;
+            $totMonths += $since_start->m;
+        } catch (\Exception $e) {
+            error_log($e);
+        }
+
+        if ((is_array($user) && array_key_exists('last_login', $user) && is_null($user['last_login'])) || ($totMonths >= 3)) {
+            $userInfo = TableRegistry::get('User.Users')->get($user['id']);
+            if ($userInfo->password) {
+                if(($this->request->params['action'] !== 'Accounts')){
+                    $this->Alert->warning('security.login.changePassword');
+                    $lastLogin = $userInfo->last_login;
+                    $this->request->session()->write('Auth.User.last_login', $lastLogin);
+                    $this->redirect(['plugin' => 'Profile', 'controller' => 'Profiles', 'action' => 'Accounts', 'edit', $this->ControllerAction->paramsEncode(['id' => $user['id']])]);
+    
+                }
+            }
+        }
+        $header = __('Home Page');
+        $this->set('contentHeader', $header);
         }
     }
