+++ src/Controller/DashboardController.php
@@ -42,35 +42,6 @@ class DashboardController extends AppController
         return true;
     }
 
-    public function beforeFilter(Event $event)
-    {
-        parent::beforeFilter($event);
-        $user = $this->Auth->user();
-        $userData = TableRegistry::get("security_users")->get($user['id']);
-        try {
-            date_default_timezone_set('Asia/Colombo');
-            $start_date = new DateTime($userData->security_timeout);
-            $since_start = $start_date->diff(new DateTime(date('Y-m-d h:i:s a', time())));
-            $totMonths = $since_start->y *12;
-            $totMonths += $since_start->m;
-        } catch (\Exception $e) {
-            error_log($e);
-        }
-
-        if ((is_array($user) && array_key_exists('last_login', $user) && is_null($user['last_login'])) || ($totMonths >= 3)) {
-            $userInfo = TableRegistry::get('User.Users')->get($user['id']);
-            if ($userInfo->password) {
-                $this->Alert->warning('security.login.changePassword');
-                $lastLogin = $userInfo->last_login;
-                $this->request->session()->write('Auth.User.last_login', $lastLogin);
-                $this->redirect(['plugin' => 'Profile', 'controller' => 'Profiles', 'action' => 'Accounts', 'edit', $this->ControllerAction->paramsEncode(['id' => $user['id']])]);
-            }
-
-        }
-        $header = __('Home Page');
-        $this->set('contentHeader', $header);
-    }
-
     public function onInitialize(Event $event, Table $model, ArrayObject $extra)
     {
         // set header
